requirements.txt
tenacity~=8.1.0
python-dotenv~=0.21.0
matplotlib~=3.6.2
seaborn~=0.12.1
langchain~=0.0.73
retry~=0.9.2
setuptools~=60.2.0
gensim~=4.3.0
tabulate
transformers
ml_collections
fschat>=0.2.20
huggingface_hub==0.16.4
numpy
nltk==3.8.1
openai==0.27.8
scikit-learn==1.2.2
sentence-transformers==2.2.2
tokenizers==0.13.3
wandb==0.14.0
tqdm==4.65.0
pandas==2.0.3
accelerate==0.21.0
protobuf==3.20
datasets==2.14.3
